Started modifying demo part1
